Implement methods as traits for portability across multiple API clients
